Refactor to stop passing references to iframes around
